ed25519: add back `serde_bytes` support
Since `serdect` changes the wire format, this should at least provide
backwards compatibility for `serde_bytes` users.
