+++ ed25519/Cargo.toml
@@ -23,6 +23,7 @@ signature = { version = "3", default-features = false }
 # optional dependencies
 pkcs8 = { version = "0.11", optional = true }
 serdect = { version = "0.4", optional = true, default-features = false }
+serde_bytes = { version = "0.11", optional = true, default-features = false }
 zeroize = { version = "1", optional = true, default-features = false }
 zerocopy = { version = "0.8", optional = true, features = ["derive"] }
 
@@ -36,6 +37,7 @@ default = ["alloc"]
 alloc = ["pkcs8?/alloc", "signature/alloc"]
 pem = ["alloc", "pkcs8/pem"]
 serde = ["dep:serdect"]
+serde_bytes = ["serde", "dep:serde_bytes"]
 
 [lints]
 workspace = true
